refactor: ExtensionInstallCheckDialog 분리
