+++ packages/web/src/app/[lng]/(auth)/memos/_components/index.ts
@@ -5,3 +5,4 @@ export { default as MemoView } from './MemoView';
 export { default as MemoSearchFormProvider } from './MemoSearchFormProvider';
 export { default as MemoSearchForm } from './MemoSearchForm';
 export { default as InitSentryUserInfo } from './InitSentryUserInfo';
+export { default as ExtensionInstallCheckDialog } from './ExtensionInstallCheckDialog';
